UI refactored
