+++ Cargo.toml
@@ -1,4 +1,4 @@
-workspace = { members = [ "src/app_props", "src/db","src/file_system", "src/img_azure", "src/trie", "src/vectorization"] }
+workspace = { members = [ "src/app_props", "src/db","src/file_system", "src/img_azure", "src/arc_str", "src/vectorization", "src/ui_facade"] }
 [package]
 name = "file-search"
 version = "0.1.0"
@@ -8,19 +8,13 @@ edition = "2021"
 
 [dependencies]
 file_system = { path = "src/file_system" }
-num_cpus = "1.13.0"
-trie = { path = "src/trie" }
-serde = { version = "1.0", features = ["derive"] }
-bincode = "1.3.3"
+
 dioxus = { version = "0.4.0" }
 dioxus-desktop = { version = "0.4.0" }
-tokio = { version = "1.37.0", features = ["rt", "rt-multi-thread", "macros"] }
 serde_json = "1.0.115"
 trie-rs = "0.2.0"
-once_cell = "1.19.0"
-rayon = "1.5.1"
-db = { path = "src/db" }
-app_props = { path = "src/app_props" }
-img_azure = { path = "src/img_azure" }
-vectorization = { path = "src/vectorization" }
+tokio = { version = "1.37.0", features = ["rt", "rt-multi-thread", "macros"] }
+
 manganis = "0.2.2"
+ui_facade = { path = "src/ui_facade" }
+app_props = { path = "src/app_props" }
