+++ src/app_props/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-trie = {path = "../trie" }
+arc_str = {path = "../arc_str" }
 trie-rs = "0.2.0"
 bincode = "1.3.3"
 vectorization = { path = "../vectorization" }
