+++ src/arc_str/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "trie"
+name = "arc_str"
 version = "0.1.0"
 edition = "2021"
