+++ src/arc_str/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod arc_str;
