--- src/trie/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod node;
-pub mod trie;
-pub mod arc_str;
